Remove redundant code for element pv creation
The bend elements had their own code to create their pvs. Now they use the same code as for regular elements. I had to refactor the function a little bit to keep the same behaviour as before. This being that only 1 pv is created for all bend elements and each new element adds its index to a list stored in this pv