_in_records_ now always contains a list of indexes
Previously there was either a single index or a list in special cases, now we always use a list. During creation we check that it is a list. Now we can assume it is a list after it has been created. I also made the update_pvs function loop through the indexes rather than just update index[0]