Handle updating _record_names better
We now update this after each group of pvs is created, so that when we start creating the next group of pvs, it is up to date as it is needed during creation of some record types. I also improved logging in atip_ioc_entry and reduced a caget timeout. Virtac is now starting up in ~6-7 seconds.